+++ Reporting/Compliance/CB Configuration Baseline Compliance/CB Configuration Baseline Compliance by Company with Values.sql
@@ -5,10 +5,8 @@
 * Modified by       |    Date    | Revision | Comments                                                  *
 * _____________________________________________________________________________________________________ *
 * Ioan Popovici     | 2017-09-22 | v1.0     | First version                                             *
-* Ioan Popovici     | 2018-01-17 | v1.1     | Fixed: Actual Value is NULL compliance is not displayed   *
-* Ioan Popovici     | 2018-06-21 | v2.0     | Completely re-written to optimize speed                   *
 * ===================================================================================================== *
-*                                                                                                       *
+* Release info moved separate markdown file, see notes.                                                 *
 *********************************************************************************************************
 
 .SYNOPSIS
@@ -16,6 +14,8 @@
 .DESCRIPTION
     This SQL Query is used to get the Compliance by Company and Actual Values of a Configuration Baseline Result.
 .NOTES
+    Release notes
+        https://github.com/JhonnyTerminus/SCCMZone/blob/master/Reporting/Compliance/CB%20Configuration%20Baseline%20Compliance/CHANGELOG.md
     Part of a report should not be run separately.
 .LINK
     https://SCCM-Zone.com
@@ -30,74 +30,77 @@
 /* Remove previous temporary table if exists */
 IF OBJECT_ID (N'TempDB.DBO.#CIComplianceStatusDetails') IS NOT NULL
     BEGIN
-        DROP TABLE #CIComplianceStatusDetails;
-    END;
+    DROP TABLE #CIComplianceStatusDetails;
+END;
 
-/* Get Configuration Item Current Value */
-WITH CTE
-    AS (
-        SELECT
+/* Get configuration item current value for collection members and insert the result in a temporary table */
+WITH
+    CTE
+    AS
+    (
+        SELECT DISTINCT
             CIComplianceStatusDetails.ResourceID,
+            CIComplianceStatusDetails.CIVersion AS CBRevision,
             CIComplianceStatusDetails.CurrentValue,
             CIComplianceStatusDetails.LastComplianceMessageTime,
-            RN = ROW_NUMBER()
-            OVER (
-                PARTITION BY CIComplianceStatusDetails.Netbios_Name0 ORDER BY CIComplianceStatusDetails.LastComplianceMessageTime DESC
-            )
+            CIComplianceStatusDetails.Netbios_Name0
         FROM dbo.fn_rbac_CICurrentSettingsComplianceStatusDetail(@UserSIDs) AS CIComplianceStatusDetails
         WHERE CIComplianceStatusDetails.CI_ID
             IN (
                 SELECT ReferencedCI_ID
-                FROM dbo.fn_rbac_CIRelation_All(@UserSIDs)
-                WHERE CI_ID = @BaselineID
-                    AND RelationType NOT IN ('7', '0') --Exlude itself and no relation
+        FROM dbo.fn_rbac_CIRelation_All(@UserSIDs)
+        WHERE CI_ID = @BaselineID
+            AND RelationType NOT IN ('7', '0') --Exlude itself and no relation
             )
     )
 SELECT
     ResourceID,
+    CBRevision,
     CurrentValue,
     LastComplianceMessageTime
 INTO #CIComplianceStatusDetails
 FROM CTE
-WHERE RN = 1
 ORDER BY ResourceID;
 
+/* Get the other details and join with them with the temporary table based on ResourceID */
 SELECT
 
     /* IMPORTANT! YOU NEED TO ENABLE THE COMPANY FIELD GATHERING FOR SYSTEM DISCOVERY OTHERWISE THIS COLUMN IS NOT AVAILABLE */
-    Users.Company0 AS 'Company',
+    ComputerSystem.Company0 AS 'Company',
 
     /* CUSTOM FUNCTION LEAVE DISABLED */
-    --( SELECT [CM_Tools].[dbo].[ufn_GetCompany_by_ResourceID]([Users].[ResourceID]) ) AS [Company],
+    --( SELECT [CM_Tools].[dbo].[ufn_GetCompany_by_ResourceID]([ComputerSystem].[ResourceID]) ) AS [Company],
 
+    CollectionMembers.ResourceID,
     CIComplianceState.DisplayName,
-    CIComplianceState.ComplianceStateName AS ComplianceState,
-    Computer.Name0 AS DeviceName,
-    Users.User_Name0 AS UserName,
+    ( SELECT ISNULL(CIComplianceState.ComplianceStateName, 'Unknown' )) AS ComplianceState,
     OperatingSystem.Caption0 AS OperatingSystem,
-    Computer.Model0 AS [Model],
-    CIComplianceState.CIVersion,
+    ( SELECT ISNULL(ComputerSystem.UserName0, 'Unknown' )) AS UserName,
+    ComputerSystem.Name0 AS DeviceName,
+    ComputerSystem.Model0 AS Model,
+    ComputerStatus.LastHWScan,
+    CIComplianceState.CIVersion AS CIRevision,
+    CIComplianceStatusDetails.CBRevision,
     CIComplianceStatusDetails.CurrentValue,
-    CIComplianceStatusDetails.LastComplianceMessageTime AS LastEvaluation,
-    ComputerStatus.LastHWScan AS LastHWScan
-FROM v_BaselineTargetedComputers Baseline
-    JOIN dbo.fn_rbac_R_System(@UserSIDs) AS Users ON Users.ResourceID = Baseline.ResourceID
-    JOIN v_ClientCollectionMembers AS Collections ON Collections.ResourceID = Users.ResourceID
-    JOIN v_GS_COMPUTER_SYSTEM Computer ON Computer.ResourceID = Users.ResourceID
-    JOIN v_GS_OPERATING_SYSTEM AS OperatingSystem ON OperatingSystem.ResourceID = Users.ResourceID
-    JOIN dbo.fn_rbac_CIRelation_All(@UserSIDs) AS BaselineCIs ON BaselineCIs.CI_ID = @BaselineID
-        AND BaselineCIs.RelationType NOT IN ('7', '0') --Exlude itself and no relation
-    JOIN dbo.fn_rbac_ConfigurationItems(@UserSIDs) CIInfo ON CIInfo.CI_ID = BaselineCIs.ReferencedCI_ID
-    JOIN dbo.fn_rbac_ListCI_ComplianceState(@LocaleID, @UserSIDs) AS CIComplianceState ON CIComplianceState.CI_ID = BaselineCIs.ReferencedCI_ID
-        AND CIComplianceState.ResourceID = Users.ResourceID
-    JOIN v_GS_WORKSTATION_STATUS AS ComputerStatus ON ComputerStatus.ResourceID = Users.ResourceID
-    LEFT OUTER JOIN #CIComplianceStatusDetails AS CIComplianceStatusDetails ON CIComplianceStatusDetails.ResourceID = Baseline.ResourceID
-WHERE Collections.CollectionID = @CollectionID
-    AND Baseline.CI_ID = @BaselineID
+    CIComplianceStatusDetails.LastComplianceMessageTime,
+    CIComplianceStatusDetails.LastComplianceMessageTime AS LastEvaluation
+FROM
+    v_ClientCollectionMembers AS CollectionMembers
+    LEFT JOIN v_GS_OPERATING_SYSTEM AS OperatingSystem ON OperatingSystem.ResourceID = CollectionMembers.ResourceID
+    LEFT JOIN v_GS_COMPUTER_SYSTEM AS ComputerSystem ON ComputerSystem.ResourceID = CollectionMembers.ResourceID
+    LEFT JOIN v_GS_WORKSTATION_STATUS AS ComputerStatus ON ComputerStatus.ResourceID = CollectionMembers.ResourceID
+    FULL JOIN dbo.fn_rbac_ListCI_ComplianceState(@LocaleID, @UserSIDs) AS CIComplianceState ON CIComplianceState.ResourceID = CollectionMembers.ResourceID
+        AND CIComplianceState.CI_ID = @BaselineID
+    FULL JOIN #CIComplianceStatusDetails AS CIComplianceStatusDetails ON CIComplianceStatusDetails.ResourceID = CollectionMembers.ResourceID
+WHERE
+	CollectionMembers.CollectionID = @CollectionID
 ORDER BY
-	Company,
+    Company,
 	DisplayName,
 	ComplianceStateName
+
+/* Remove  temporary table */
+DROP TABLE #CIComplianceStatusDetails;
 
 /* #endregion */
 /*##=============================================*/